requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
langserve
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
absl-py
sentence_transformers
langchain_huggingface
faiss-cpu
protobuf==3.20.3
langchain_chroma
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
openai
arxiv
wikipedia
duckduckgo-search